이전 버전 v2

+++ src/main/kotlin/com/yoong/myissue/domain/team/service/TeamServiceImpl.kt
@@ -1,6 +1,6 @@
 package com.yoong.myissue.domain.team.service
 
-import com.yoong.myissue.common.annotationGather.*
+import com.yoong.myissue.common.clazz.ValidAuthentication
 import com.yoong.myissue.common.enumGather.AuthenticationType
 import com.yoong.myissue.domain.issue.enumGather.Role
 import com.yoong.myissue.domain.member.service.ExternalMemberService
@@ -11,6 +11,7 @@ import com.yoong.myissue.domain.team.repository.TeamRepository
 import com.yoong.myissue.exception.clazz.DummyTeamException
 import com.yoong.myissue.exception.clazz.DuplicatedModelException
 import com.yoong.myissue.exception.clazz.ModelNotFoundException
+import com.yoong.myissue.exception.clazz.IllegalArgumentException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -20,12 +21,16 @@ import org.springframework.transaction.annotation.Transactional
 class TeamServiceImpl(
     private val teamRepository: TeamRepository,
     private val memberService: ExternalMemberService,
-    private val externalTeamService: ExternalTeamService
-): TeamService {
+    private val externalTeamService: ExternalTeamService,
+    private val validAuthentication: ValidAuthentication
+){
 
-    @CheckAuthentication(authenticationType = AuthenticationType.USER)
-    override fun createTeam(email: String, createTeamRequest: TeamRequest): String {
+    fun createTeam(email: String, createTeamRequest: TeamRequest): String {
+
         val member = memberService.searchEmail(email)
+
+        validAuthentication.role(member.getRole(), AuthenticationType.USER)
+
 
         if(teamRepository.existsByName(createTeamRequest.name)) throw DuplicatedModelException("팀 명", createTeamRequest.name)
 
@@ -38,25 +43,39 @@ class TeamServiceImpl(
         return "팀이 생성 되었습니다 팀 명 : ${createTeamRequest.name}"
     }
 
-    @CheckAuthentication(authenticationType = AuthenticationType.LEADER_AND_ADMIN)
-    @CheckDummyTeam
     @Transactional(readOnly = true)
-    override fun getTeamById(email: String, teamId: Long): TeamResponse {
+    fun getTeamById(email: String, teamId: Long): TeamResponse {
+
+        val member = memberService.searchEmail(email)
+
+        validAuthentication.role(member.getRole(), AuthenticationType.LEADER_AND_ADMIN)
+
+        if (teamId == DUMMY_TEAM) throw DummyTeamException(null)
+
         return teamRepository.findByIdOrNull(teamId)?.toTeamResponse() ?: throw ModelNotFoundException("id", teamId.toString())
     }
 
-    @CheckAuthentication(authenticationType = AuthenticationType.ADMIN)
     @Transactional(readOnly = true)
-    override fun getTeamList(email: String): List<TeamResponse> {
+    fun getTeamList(email: String): List<TeamResponse> {
+
+        val member = memberService.searchEmail(email)
+
+        validAuthentication.role(member.getRole(), AuthenticationType.ADMIN)
 
         return teamRepository.findAll().map { it.toTeamResponse() }
     }
 
-    @CheckAuthentication(authenticationType = AuthenticationType.LEADER_AND_ADMIN)
-    @LeaderChoosesOtherTeam
-    @CheckDummyTeam
-    override fun updateTeam(email: String, teamId: Long, teamRequest: TeamRequest): String {
+    fun updateTeam(email: String, teamId: Long, teamRequest: TeamRequest): String {
+
+        val member = memberService.searchEmail(email)
+
+        validAuthentication.role(member.getRole(), AuthenticationType.LEADER_AND_ADMIN)
+
         val team = teamRepository.findByIdOrNull(teamId) ?: throw ModelNotFoundException("id", teamId.toString())
+
+        if (teamId == DUMMY_TEAM) throw DummyTeamException(null)
+
+        if(member.getRole() == Role.LEADER && team != member.getTeam()) throw IllegalArgumentException("본인 소속 외에 다른 팀은 선택할 수 없습니다")
 
         team.update(teamRequest)
 
@@ -65,25 +84,34 @@ class TeamServiceImpl(
         return "팀 명 변경이 완료 되었습니다 변경된 팀 명 : ${teamRequest.name}"
     }
 
-    @CheckAuthentication(authenticationType = AuthenticationType.LEADER_AND_ADMIN)
-    @LeaderChoosesOtherTeam
-    @CheckDummyTeam
-    override fun deleteTeam(email: String, teamId: Long): String {
+    fun deleteTeam(email: String, teamId: Long): String {
+
+        val member = memberService.searchEmail(email)
+
+        validAuthentication.role(member.getRole(), AuthenticationType.LEADER_AND_ADMIN)
 
         val team = teamRepository.findByIdOrNull(teamId) ?: throw ModelNotFoundException("id", teamId.toString())
+
+        if (teamId == DUMMY_TEAM) throw DummyTeamException(null)
+
+        if(member.getRole() == Role.LEADER && team != member.getTeam()) throw IllegalArgumentException("본인 소속 외에 다른 팀은 선택할 수 없습니다")
 
         teamRepository.delete(team)
 
         return "삭제가 완료 되었습니다"
     }
 
-    @CheckAuthentication(authenticationType = AuthenticationType.LEADER)
-    @CheckUser
-    @CheckMine
-    override fun inviteTeam(email: String, memberId: Long): String {
+    fun inviteTeam(email: String, memberId: Long): String {
+
+        val leader = memberService.searchEmail(email)
+
+        validAuthentication.role(leader.getRole(), AuthenticationType.LEADER_AND_ADMIN)
+
+        if(leader.getRole() != Role.USER) throw IllegalArgumentException("리더 및 관리자는 팀원으로 추가 및 방출할 수 없습니다")
 
         val member = memberService.searchId(memberId)
-        val leader = memberService.searchEmail(email)
+
+        if(member == leader) throw DuplicatedModelException("맴버", "나 자신을 선택할 수 없습니다")
 
         when(member.getTeam().getId()){
             DUMMY_TEAM -> member.updateTeam(leader.getTeam())
@@ -94,12 +122,17 @@ class TeamServiceImpl(
         return "새로운 팀 원이 등록 되었습니다"
     }
 
-    @CheckAuthentication(authenticationType = AuthenticationType.ADMIN)
-    @CheckUser
-    @CheckMine
-    override fun inviteMemberByAdmin(email: String, memberId: Long, teamId: Long): String {
+    fun inviteMemberByAdmin(email: String, memberId: Long, teamId: Long): String {
+
+        val admin = memberService.searchEmail(email)
+
+        validAuthentication.role(admin.getRole(), AuthenticationType.ADMIN)
 
         val member = memberService.searchId(memberId)
+
+        if(member.getRole() != Role.USER) throw IllegalArgumentException("리더 및 관리자는 팀원으로 추가 및 방출할 수 없습니다")
+
+        if(member == admin) throw DuplicatedModelException("맴버", "나 자신을 선택할 수 없습니다")
 
         val team = externalTeamService.getTeamById(teamId)
 
@@ -113,14 +146,17 @@ class TeamServiceImpl(
         return "새로운 팀 원이 ${team.getTeamName()} 팀에 등록 되었습니다"
     }
 
-    @CheckAuthentication(authenticationType = AuthenticationType.LEADER_AND_ADMIN)
-    @CheckUser
-    @CheckMine
-    override fun firedMember(email: String, memberId: Long): String? {
+    fun firedMember(email: String, memberId: Long): String? {
+
+        val leader = memberService.searchEmail(email)
+
+        validAuthentication.role(leader.getRole(), AuthenticationType.LEADER_AND_ADMIN)
 
         val dummyTeam = externalTeamService.getDummyTeam()
         val member = memberService.searchId(memberId)
-        val leader = memberService.searchEmail(email)
+        if(member.getRole() != Role.USER) throw IllegalArgumentException("리더 및 관리자는 팀원으로 추가 및 방출할 수 없습니다")
+        if(member == leader) throw DuplicatedModelException("맴버", "나 자신을 선택할 수 없습니다")
+
 
         if(leader.getRole() == Role.LEADER){
             when(member.getTeam()){
